Fix build from stlport.
This will need to be submitted upstream, but for now it's blocking the
build.

Can't test yet as the build is still red, but there's no risk doing
this for now because none of these modules are used yet.

Bug: 20915699
Change-Id: I9e7c8e237bec134985735dee4b4225b29793a792
(cherry picked from commit 6368380f06a48b0049118140f6ac5059f4e7f0f0)