Use the shared library VNDK version of libxml2
Previously the parameter-framework was using the static version of
libxml2. This meant that the libxml2 dependency libicu had to be listed
as a parameter-framework dependency. Nevertheless libicu is not vendor
accessible.

As libxml2 is part of the VNDK, use its shared-library variant allows to
remove the explicit libicu dependency as well as reducing code size and
allows independent VNDK update.

Bug: 130284799
Test: make

Change-Id: I95b6a72c72161c8267786790de7093d8a641c392
Signed-off-by: Kevin Rocard <krocard@google.com>