Decorrelate tuning mode and tuning tool
In order to set the configuration at runtime (required to migrate
Setting from Configurable AudioPolicy Engine to AudioHAL), it is
required to use the full connector and its feature programmatically.
The tuning mode must be set to add settings.

However, this guard is not only applicable when using tuning tool or
when using the API programmatically.

This CL moves the guard from the tuning API to the tunning command API.

Bug: 307310023
Test: manual: ls /dev/socket/audioserver/ shall be empty in user build

Change-Id: I3e6377551eb745b2adf60cac04e4bcc6383b33f3
Signed-off-by: François Gaffie <francois.gaffie@renault.com>